WIP: rewrite `#[storage]`
