+++ crates/svm-sdk/macros/src/function/endpoint.rs
@@ -4,11 +4,11 @@ use quote::{quote, ToTokens};
 use syn::{Error, FnArg, Pat, PatType, Result, ReturnType, Type};
 
 use super::attr;
-use attr::{has_endpoint_attr, has_fundable_attr, FuncAttribute};
+use attr::{has_endpoint_attr, has_fundable_attr, FuncAttr};
 
 use crate::Function;
 
-pub fn expand(func: &Function, attrs: &[FuncAttribute]) -> Result<TokenStream> {
+pub fn expand(func: &Function, attrs: &[FuncAttr]) -> Result<TokenStream> {
     debug_assert!(has_endpoint_attr(attrs));
 
     validate_sig(func)?;
